Recommit of 12163 to the trunk.
The fit_copy macro checks for the start address being off the top of the
screen, and clips it to zero. When it does this, it does: data -= y * raster,
which gives problems if raster is a uint ( as uint * int == uint in C) and
data is a 64 bit pointer.

The fix is simply to cast the result to an int before using it. This solves
various SEGVs with the mono_copy_mono using mono_copy_rop code.

No cluster differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12179 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/base/gxdevice.h
@@ -1,6 +1,6 @@
 /* Copyright (C) 2001-2006 Artifex Software, Inc.
    All Rights Reserved.
-  
+
    This software is provided AS-IS with no warranty, either express or
    implied.
 
@@ -17,7 +17,7 @@
 #ifndef gxdevice_INCLUDED
 #  define gxdevice_INCLUDED
 
-#include "stdio_.h"		/* for FILE */
+#include "stdio_.h"             /* for FILE */
 #include "gxdevcli.h"
 #include "gsfname.h"
 #include "gsparam.h"
@@ -81,30 +81,30 @@
  * the next element of the structure.
  */
 #define std_device_part1_(devtype, ptr_procs, dev_name, stype, open_init)\
-	sizeof(devtype), ptr_procs, dev_name,\
-	0 /*memory*/, stype, 0 /*stype_is_dynamic*/, 0 /*finalize*/,\
-	{ 0 } /*rc*/, 0 /*retained*/, open_init() /*is_open, max_fill_band*/
-	/* color_info goes here */
+        sizeof(devtype), ptr_procs, dev_name,\
+        0 /*memory*/, stype, 0 /*stype_is_dynamic*/, 0 /*finalize*/,\
+        { 0 } /*rc*/, 0 /*retained*/, open_init() /*is_open, max_fill_band*/
+        /* color_info goes here */
 /*
  * The MetroWerks compiler has some bizarre bug that produces a spurious
  * error message if the width and/or height are defined as 0 below,
  * unless we use the +/- workaround in the next macro.
  */
 #define std_device_part2_(width, height, x_dpi, y_dpi)\
-	{ gx_no_color_index, gx_no_color_index },\
+        { gx_no_color_index, gx_no_color_index },\
           width, height, 0/*TrayOrientation*/,\
-	{ (float)((((width) * 72.0 + 0.5) - 0.5) / (x_dpi))/*MediaSize[0]*/,\
-	  (float)((((height) * 72.0 + 0.5) - 0.5) / (y_dpi))/*MediaSize[1]*/},\
-	{ 0, 0, 0, 0 }/*ImagingBBox*/, 0/*ImagingBBox_set*/,\
-	{ x_dpi, y_dpi }/*HWResolution*/, { x_dpi, y_dpi }/*MarginsHWResolution*/
+        { (float)((((width) * 72.0 + 0.5) - 0.5) / (x_dpi))/*MediaSize[0]*/,\
+          (float)((((height) * 72.0 + 0.5) - 0.5) / (y_dpi))/*MediaSize[1]*/},\
+        { 0, 0, 0, 0 }/*ImagingBBox*/, 0/*ImagingBBox_set*/,\
+        { x_dpi, y_dpi }/*HWResolution*/, { x_dpi, y_dpi }/*MarginsHWResolution*/
 
 /* offsets and margins go here */
 #define std_device_part3_()\
-	0/*PageCount*/, 0/*ShowpageCount*/, 1/*NumCopies*/, 0/*NumCopies_set*/,\
-	0/*IgnoreNumCopies*/, 0/*UseCIEColor*/, 0/*LockSafetyParams*/,\
-	0/*band_offset_x*/, 0/*band_offset_y*/, {false}/* sgr */, 0/* MaxPatternBitmap */,\
-	0/*Profile*/,\
-	{ gx_default_install, gx_default_begin_page, gx_default_end_page }
+        0/*PageCount*/, 0/*ShowpageCount*/, 1/*NumCopies*/, 0/*NumCopies_set*/,\
+        0/*IgnoreNumCopies*/, 0/*UseCIEColor*/, 0/*LockSafetyParams*/,\
+        0/*band_offset_x*/, 0/*band_offset_y*/, {false}/* sgr */, 0/* MaxPatternBitmap */,\
+        0/*Profile*/,\
+        { gx_default_install, gx_default_begin_page, gx_default_end_page }
 /*
  * We need a number of different variants of the std_device_ macro simply
  * because we can't pass the color_info or offsets/margins
@@ -118,94 +118,94 @@
  *        6 for margins/offset).
  *      ...std_color_full_body, which takes 7 values (1 for dci_std_color,
  *        6 for margins/offset).
- *      
+ *
  */
 #define std_device_body_with_macros_(dtype, pprocs, dname, stype, w, h, xdpi, ydpi, open_init, dci_macro, margins_macro)\
-	std_device_part1_(dtype, pprocs, dname, stype, open_init),\
-	dci_macro(),\
-	std_device_part2_(w, h, xdpi, ydpi),\
-	margins_macro(),\
-	std_device_part3_()
+        std_device_part1_(dtype, pprocs, dname, stype, open_init),\
+        dci_macro(),\
+        std_device_part2_(w, h, xdpi, ydpi),\
+        margins_macro(),\
+        std_device_part3_()
 
 #define std_device_std_body_type(dtype, pprocs, dname, stype, w, h, xdpi, ydpi)\
-	std_device_body_with_macros_(dtype, pprocs, dname, stype,\
-	  w, h, xdpi, ydpi,\
-	  open_init_closed, dci_black_and_white_, no_margins_)
+        std_device_body_with_macros_(dtype, pprocs, dname, stype,\
+          w, h, xdpi, ydpi,\
+          open_init_closed, dci_black_and_white_, no_margins_)
 
 #define std_device_std_body(dtype, pprocs, dname, w, h, xdpi, ydpi)\
-	std_device_std_body_type(dtype, pprocs, dname, 0, w, h, xdpi, ydpi)
+        std_device_std_body_type(dtype, pprocs, dname, 0, w, h, xdpi, ydpi)
 
 #define std_device_std_body_type_open(dtype, pprocs, dname, stype, w, h, xdpi, ydpi)\
-	std_device_body_with_macros_(dtype, pprocs, dname, stype,\
-	  w, h, xdpi, ydpi,\
-	  open_init_open, dci_black_and_white_, no_margins_)
+        std_device_body_with_macros_(dtype, pprocs, dname, stype,\
+          w, h, xdpi, ydpi,\
+          open_init_open, dci_black_and_white_, no_margins_)
 
 #define std_device_std_body_open(dtype, pprocs, dname, w, h, xdpi, ydpi)\
-	std_device_std_body_type_open(dtype, pprocs, dname, 0, w, h, xdpi, ydpi)
+        std_device_std_body_type_open(dtype, pprocs, dname, 0, w, h, xdpi, ydpi)
 
 #define std_device_full_body_type(dtype, pprocs, dname, stype, w, h, xdpi, ydpi, ncomp, depth, mg, mc, dg, dc, xoff, yoff, lm, bm, rm, tm)\
-	std_device_part1_(dtype, pprocs, dname, stype, open_init_closed),\
-	dci_values(ncomp, depth, mg, mc, dg, dc),\
-	std_device_part2_(w, h, xdpi, ydpi),\
-	offset_margin_values(xoff, yoff, lm, bm, rm, tm),\
-	std_device_part3_()
+        std_device_part1_(dtype, pprocs, dname, stype, open_init_closed),\
+        dci_values(ncomp, depth, mg, mc, dg, dc),\
+        std_device_part2_(w, h, xdpi, ydpi),\
+        offset_margin_values(xoff, yoff, lm, bm, rm, tm),\
+        std_device_part3_()
 
 #define std_device_full_body_type_extended(dtype, pprocs, dname, stype, w, h, xdpi, ydpi, mcomp, ncomp, pol, depth, gi, mg, mc, dg, dc, ef, cn, xoff, yoff, lm, bm, rm, tm)\
-	std_device_part1_(dtype, pprocs, dname, stype, open_init_closed),\
+        std_device_part1_(dtype, pprocs, dname, stype, open_init_closed),\
         dci_extended_alpha_values(mcomp, ncomp, pol, depth, gi, mg, mc, dg, dc, 1, 1, ef, cn), \
-	std_device_part2_(w, h, xdpi, ydpi),\
-	offset_margin_values(xoff, yoff, lm, bm, rm, tm),\
-	std_device_part3_()
+        std_device_part2_(w, h, xdpi, ydpi),\
+        offset_margin_values(xoff, yoff, lm, bm, rm, tm),\
+        std_device_part3_()
 
 #define std_device_full_body(dtype, pprocs, dname, w, h, xdpi, ydpi, ncomp, depth, mg, mc, dg, dc, xoff, yoff, lm, bm, rm, tm)\
-	std_device_full_body_type(dtype, pprocs, dname, 0, w, h, xdpi, ydpi,\
-	    ncomp, depth, mg, mc, dg, dc, xoff, yoff, lm, bm, rm, tm)
+        std_device_full_body_type(dtype, pprocs, dname, 0, w, h, xdpi, ydpi,\
+            ncomp, depth, mg, mc, dg, dc, xoff, yoff, lm, bm, rm, tm)
 
 #define std_device_dci_alpha_type_body(dtype, pprocs, dname, stype, w, h, xdpi, ydpi, ncomp, depth, mg, mc, dg, dc, ta, ga)\
-	std_device_part1_(dtype, pprocs, dname, stype, open_init_closed),\
-	dci_alpha_values(ncomp, depth, mg, mc, dg, dc, ta, ga),\
-	std_device_part2_(w, h, xdpi, ydpi),\
-	offset_margin_values(0, 0, 0, 0, 0, 0),\
-	std_device_part3_()
+        std_device_part1_(dtype, pprocs, dname, stype, open_init_closed),\
+        dci_alpha_values(ncomp, depth, mg, mc, dg, dc, ta, ga),\
+        std_device_part2_(w, h, xdpi, ydpi),\
+        offset_margin_values(0, 0, 0, 0, 0, 0),\
+        std_device_part3_()
 
 #define std_device_dci_type_body(dtype, pprocs, dname, stype, w, h, xdpi, ydpi, ncomp, depth, mg, mc, dg, dc)\
-	std_device_dci_alpha_type_body(dtype, pprocs, dname, stype, w, h,\
-	  xdpi, ydpi, ncomp, depth, mg, mc, dg, dc, 1, 1)
+        std_device_dci_alpha_type_body(dtype, pprocs, dname, stype, w, h,\
+          xdpi, ydpi, ncomp, depth, mg, mc, dg, dc, 1, 1)
 
 #define std_device_dci_body(dtype, pprocs, dname, w, h, xdpi, ydpi, ncomp, depth, mg, mc, dg, dc)\
-	std_device_dci_type_body(dtype, pprocs, dname, 0,\
-	  w, h, xdpi, ydpi, ncomp, depth, mg, mc, dg, dc)
+        std_device_dci_type_body(dtype, pprocs, dname, 0,\
+          w, h, xdpi, ydpi, ncomp, depth, mg, mc, dg, dc)
 
 #define std_device_color_full_body(dtype, pprocs, dname, w, h, xdpi, ydpi, depth, max_value, dither, xoff, yoff, lm, bm, rm, tm)\
-	std_device_part1_(dtype, pprocs, dname, 0, open_init_closed),\
-	dci_color(depth, max_value, dither),\
-	std_device_part2_(w, h, xdpi, ydpi),\
-	offset_margin_values(xoff, yoff, lm, bm, rm, tm),\
-	std_device_part3_()
+        std_device_part1_(dtype, pprocs, dname, 0, open_init_closed),\
+        dci_color(depth, max_value, dither),\
+        std_device_part2_(w, h, xdpi, ydpi),\
+        offset_margin_values(xoff, yoff, lm, bm, rm, tm),\
+        std_device_part3_()
 
 #define std_device_color_body(dtype, pprocs, dname, w, h, xdpi, ydpi, depth, max_value, dither)\
-	std_device_color_full_body(dtype, pprocs, dname,\
-	  w, h, xdpi, ydpi,\
-	  depth, max_value, dither,\
-	  0, 0, 0, 0, 0, 0)
+        std_device_color_full_body(dtype, pprocs, dname,\
+          w, h, xdpi, ydpi,\
+          depth, max_value, dither,\
+          0, 0, 0, 0, 0, 0)
 
 #define std_device_color_stype_body(dtype, pprocs, dname, stype, w, h, xdpi, ydpi, depth, max_value, dither)\
-	std_device_part1_(dtype, pprocs, dname, stype, open_init_closed),\
-	dci_color(depth, max_value, dither),\
-	std_device_part2_(w, h, xdpi, ydpi),\
-	offset_margin_values(0, 0, 0, 0, 0, 0),\
-	std_device_part3_()
+        std_device_part1_(dtype, pprocs, dname, stype, open_init_closed),\
+        dci_color(depth, max_value, dither),\
+        std_device_part2_(w, h, xdpi, ydpi),\
+        offset_margin_values(0, 0, 0, 0, 0, 0),\
+        std_device_part3_()
 
 #define std_device_std_color_full_body_type(dtype, pprocs, dname, stype, w, h, xdpi, ydpi, depth, xoff, yoff, lm, bm, rm, tm)\
-	std_device_part1_(dtype, pprocs, dname, stype, open_init_closed),\
-	dci_std_color(depth),\
-	std_device_part2_(w, h, xdpi, ydpi),\
-	offset_margin_values(xoff, yoff, lm, bm, rm, tm),\
-	std_device_part3_()
+        std_device_part1_(dtype, pprocs, dname, stype, open_init_closed),\
+        dci_std_color(depth),\
+        std_device_part2_(w, h, xdpi, ydpi),\
+        offset_margin_values(xoff, yoff, lm, bm, rm, tm),\
+        std_device_part3_()
 
 #define std_device_std_color_full_body(dtype, pprocs, dname, w, h, xdpi, ydpi, depth, xoff, yoff, lm, bm, rm, tm)\
-	std_device_std_color_full_body_type(dtype, pprocs, dname, 0,\
-	    w, h, xdpi, ydpi, depth, xoff, yoff, lm, bm, rm, tm)
+        std_device_std_color_full_body_type(dtype, pprocs, dname, 0,\
+            w, h, xdpi, ydpi, depth, xoff, yoff, lm, bm, rm, tm)
 
 /* ---------------- Default implementations ---------------- */
 
@@ -225,7 +225,7 @@ dev_proc_tile_rectangle(gx_default_tile_rectangle);
 dev_proc_copy_mono(gx_default_copy_mono);
 dev_proc_copy_color(gx_default_copy_color);
 dev_proc_draw_line(gx_default_draw_line);
-dev_proc_get_bits(gx_no_get_bits);	/* gives error */
+dev_proc_get_bits(gx_no_get_bits);      /* gives error */
 dev_proc_get_bits(gx_default_get_bits);
 dev_proc_get_params(gx_default_get_params);
 dev_proc_put_params(gx_default_put_params);
@@ -233,13 +233,13 @@ dev_proc_map_cmyk_color(gx_default_map_cmyk_color);
 dev_proc_get_xfont_procs(gx_default_get_xfont_procs);
 dev_proc_get_xfont_device(gx_default_get_xfont_device);
 dev_proc_map_rgb_alpha_color(gx_default_map_rgb_alpha_color);
-dev_proc_get_page_device(gx_default_get_page_device);	/* returns NULL */
-dev_proc_get_page_device(gx_page_device_get_page_device);	/* returns dev */
+dev_proc_get_page_device(gx_default_get_page_device);   /* returns NULL */
+dev_proc_get_page_device(gx_page_device_get_page_device);       /* returns dev */
 dev_proc_get_alpha_bits(gx_default_get_alpha_bits);
-dev_proc_copy_alpha(gx_no_copy_alpha);	/* gives error */
+dev_proc_copy_alpha(gx_no_copy_alpha);  /* gives error */
 dev_proc_copy_alpha(gx_default_copy_alpha);
 dev_proc_get_band(gx_default_get_band);
-dev_proc_copy_rop(gx_no_copy_rop);	/* gives error */
+dev_proc_copy_rop(gx_no_copy_rop);      /* gives error */
 dev_proc_copy_rop(gx_default_copy_rop);
 dev_proc_fill_path(gx_default_fill_path);
 dev_proc_stroke_path(gx_default_stroke_path);
@@ -252,12 +252,12 @@ dev_proc_begin_image(gx_default_begin_image);
 dev_proc_image_data(gx_default_image_data);
 dev_proc_end_image(gx_default_end_image);
 dev_proc_strip_tile_rectangle(gx_default_strip_tile_rectangle);
-dev_proc_strip_copy_rop(gx_no_strip_copy_rop);	/* gives error */
+dev_proc_strip_copy_rop(gx_no_strip_copy_rop);  /* gives error */
 dev_proc_strip_copy_rop(gx_default_strip_copy_rop);
 dev_proc_get_clipping_box(gx_default_get_clipping_box);
 dev_proc_get_clipping_box(gx_get_largest_clipping_box);
 dev_proc_begin_typed_image(gx_default_begin_typed_image);
-dev_proc_get_bits_rectangle(gx_no_get_bits_rectangle);	/* gives error */
+dev_proc_get_bits_rectangle(gx_no_get_bits_rectangle);  /* gives error */
 dev_proc_get_bits_rectangle(gx_default_get_bits_rectangle);
 dev_proc_map_color_rgb_alpha(gx_default_map_color_rgb_alpha);
 dev_proc_create_compositor(gx_no_create_compositor);
@@ -423,20 +423,20 @@ void gx_device_copy_params(gx_device *dev, const gx_device *target);
  * is currently allowed.
  */
 int gx_parse_output_file_name(gs_parsed_file_name_t *pfn,
-			      const char **pfmt, const char *fname,
-			      uint len, gs_memory_t *memory);
+                              const char **pfmt, const char *fname,
+                              uint len, gs_memory_t *memory);
 
 /*
  * Open the output file for a device.  Note that if the file name is empty,
  * it may be replaced with the name of a scratch file.
  */
 int gx_device_open_output_file(const gx_device * dev, char *fname,
-			       bool binary, bool positionable,
-			       FILE ** pfile);
+                               bool binary, bool positionable,
+                               FILE ** pfile);
 
 /* Close the output file for a device. */
 int gx_device_close_output_file(const gx_device * dev, const char *fname,
-				FILE *file);
+                                FILE *file);
 
 /*
  * Define the number of levels for a colorant above which we do not halftone.
@@ -471,42 +471,42 @@ dev_proc_output_page(gx_finish_output_page);
  */
 #define fit_fill_xy(dev, x, y, w, h)\
   BEGIN\
-	if ( (x | y) < 0 ) {\
-	  if ( x < 0 )\
-	    w += x, x = 0;\
-	  if ( y < 0 )\
-	    h += y, y = 0;\
-	}\
+        if ( (x | y) < 0 ) {\
+          if ( x < 0 )\
+            w += x, x = 0;\
+          if ( y < 0 )\
+            h += y, y = 0;\
+        }\
   END
 #define fit_fill_y(dev, y, h)\
   BEGIN\
-	if ( y < 0 )\
-	  h += y, y = 0;\
+        if ( y < 0 )\
+          h += y, y = 0;\
   END
 #define fit_fill_w(dev, x, w)\
   BEGIN\
-	if ( w > (dev)->width - x )\
-	  w = (dev)->width - x;\
+        if ( w > (dev)->width - x )\
+          w = (dev)->width - x;\
   END
 #define fit_fill_h(dev, y, h)\
   BEGIN\
-	if ( h > (dev)->height - y )\
-	  h = (dev)->height - y;\
+        if ( h > (dev)->height - y )\
+          h = (dev)->height - y;\
   END
 #define fit_fill_xywh(dev, x, y, w, h)\
   BEGIN\
-	fit_fill_xy(dev, x, y, w, h);\
-	fit_fill_w(dev, x, w);\
-	fit_fill_h(dev, y, h);\
+        fit_fill_xy(dev, x, y, w, h);\
+        fit_fill_w(dev, x, w);\
+        fit_fill_h(dev, y, h);\
   END
 /*
  * Clip all edges, and return from the procedure if the result is empty.
  */
 #define fit_fill(dev, x, y, w, h)\
   BEGIN\
-	fit_fill_xywh(dev, x, y, w, h);\
-	if ( w <= 0 || h <= 0 )\
-	  return 0;\
+        fit_fill_xywh(dev, x, y, w, h);\
+        if ( w <= 0 || h <= 0 )\
+          return 0;\
   END
 
 /*
@@ -520,25 +520,25 @@ dev_proc_output_page(gx_finish_output_page);
  */
 #define fit_copy_xyw(dev, data, data_x, raster, id, x, y, w, h)\
   BEGIN\
-	if ( (x | y) < 0 ) {\
-	  if ( x < 0 )\
-	    w += x, data_x -= x, x = 0;\
-	  if ( y < 0 )\
-	    h += y, data -= y * raster, id = gx_no_bitmap_id, y = 0;\
-	}\
-	if ( w > (dev)->width - x )\
-	  w = (dev)->width - x;\
+        if ( (x | y) < 0 ) {\
+          if ( x < 0 )\
+            w += x, data_x -= x, x = 0;\
+          if ( y < 0 )\
+            h += y, data -= (int)(y * raster), id = gx_no_bitmap_id, y = 0; \
+        }\
+        if ( w > (dev)->width - x )\
+          w = (dev)->width - x;\
   END
 /*
  * Clip all edges, and return from the procedure if the result is empty.
  */
 #define fit_copy(dev, data, data_x, raster, id, x, y, w, h)\
   BEGIN\
-	fit_copy_xyw(dev, data, data_x, raster, id, x, y, w, h);\
-	if ( h > (dev)->height - y )\
-	  h = (dev)->height - y;\
-	if ( w <= 0 || h <= 0 )\
-	  return 0;\
+        fit_copy_xyw(dev, data, data_x, raster, id, x, y, w, h);\
+        if ( h > (dev)->height - y )\
+          h = (dev)->height - y;\
+        if ( w <= 0 || h <= 0 )\
+          return 0;\
   END
 
 /* ---------------- Media parameters ---------------- */
@@ -547,7 +547,7 @@ dev_proc_output_page(gx_finish_output_page);
 /* The device get_params procedure would call these. */
 
 typedef struct gdev_input_media_s {
-    float PageSize[4];		/* nota bene */
+    float PageSize[4];          /* nota bene */
     const char *MediaColor;
     float MediaWeight;
     const char *MediaType;
@@ -559,13 +559,13 @@ extern const gdev_input_media_t gdev_input_media_default;
 void gdev_input_media_init(gdev_input_media_t * pim);
 
 int gdev_begin_input_media(gs_param_list * mlist, gs_param_dict * pdict,
-			   int count);
+                           int count);
 
 int gdev_write_input_page_size(int index, gs_param_dict * pdict,
-			       floatp width_points, floatp height_points);
+                               floatp width_points, floatp height_points);
 
 int gdev_write_input_media(int index, gs_param_dict * pdict,
-			   const gdev_input_media_t * pim);
+                           const gdev_input_media_t * pim);
 
 int gdev_end_input_media(gs_param_list * mlist, gs_param_dict * pdict);
 
@@ -577,10 +577,10 @@ typedef struct gdev_output_media_s {
 extern const gdev_output_media_t gdev_output_media_default;
 
 int gdev_begin_output_media(gs_param_list * mlist, gs_param_dict * pdict,
-			    int count);
+                            int count);
 
 int gdev_write_output_media(int index, gs_param_dict * pdict,
-			    const gdev_output_media_t * pom);
+                            const gdev_output_media_t * pom);
 
 int gdev_end_output_media(gs_param_list * mlist, gs_param_dict * pdict);
